modules/subscription_manager/requirements.txt
flask
flask-cors
paho-mqtt
redis
urllib3
pyyaml
prometheus_client>=0.20.0
certifi>=2024.2.2